requirements.txt
Pillow==7.0.0
pygame==2.0.0.dev6
fastapi==0.53.2
uvicorn==0.11.3
jinja2==2.11.1
aiofiles==0.4.0
